Added tooltips to Courses page
Added tooltips to the Checkmarks in Courses page, moved tooltips for Distribution Circle form in DistributionTable to DistributionsCircle
